+++ packages/pagination/src/CurrentPage.tsx
@@ -15,10 +15,6 @@ const CurrentPage = React.forwardRef<HTMLDivElement, CurrentPageProps>(
   ({ children, className, ...props }, ref) => {
     const { currentPage } = usePagination();
 
-    if (currentPage < 1) {
-      return null;
-    }
-
     const label =
       children ??
       i18n._({