+++ packages/pagination/src/PrevPage.tsx
@@ -30,7 +30,7 @@ const PrevPage = React.forwardRef<
 >(({ className, ...props }, ref) => {
   const { currentPage } = usePagination();
 
-  if (currentPage <= 1) {
+  if (currentPage === 1) {
     return null;
   }